agente_extracao/requirements.txt
streamlit==1.25.0
pandas==2.0.3
PyPDF2==3.0.1
openpyxl==3.1.2
sqlite-utils==3.35
python-dotenv==1.0.0
langchain==0.3.0
langgraph==0.2.20